+++ changelog.d/40.change
@@ -1,0 +1,1 @@
+Add a UserSessionStore and the splash screen from Element iOS.
